Move the 2 search functions in their specific file
